Corrected calculation error when changing a teams info

+++ Domain/Utilities/ServerTipOutCalculator.cs
@@ -19,7 +19,7 @@ namespace Domain.Utilities
             }
 
             decimal tipout = target * tipOutPercent;
-            return (decimal)tipout;
+            return tipout;
         }
     }
 }
